pyproject.toml
[tool.poetry]
name = "insta-spotify-status"
version = "0.1.0"
description = ""
authors = ["avilabss <contact@avilabs.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
